ci: tighten Rust clippy gate to -D warnings
rust-lint.yml now runs `cargo clippy --all-targets -- -D warnings` for the
root crate and cd_attack/, and both crates are clean under it (from ~1,250
findings on the root crate and 24 on cd_attack).

- Cargo.toml [lints] holds the whole crate-wide allow list, each with its
  reason: needless_range_loop and manual_memcpy (index loops mirror the
  specs), too_many_arguments, type_complexity, unusual_byte_groupings,
  result_unit_err (AEAD decryption fails without detail on purpose), plus
  check-cfg for `hashclash_linked` and a declared `rho_debug` feature.
- Most findings were applied with `cargo clippy --fix`; the rest were fixed
  by hand. Imports that only test modules used moved into those modules.
- Item-level #[allow] where a fix would change public API or semantics:
  large_enum_variant on CLI enums, should_implement_trait, one
  wrong_self_convention, manual_clamp where max/min maps NaN to 0, and
  dead_code on unused research helpers (kept, not deleted, since open
  branches may call them).
- Dead stores removed (slh_dsa, volcano, residual_walk, kuznyechik, ...);
  no behavior change.
- Touched files that were not rustfmt-clean are formatted, as the
  changed-file rustfmt check requires.

Validated: clippy -D warnings clean on both crates; `cargo test --release
--lib --tests` passes (2806 lib tests plus integration tests). cd_attack's
fp2_gluing_via_msolve fails identically on main (needs msolve).

+++ src/cryptanalysis/binary_isogeny.rs
@@ -233,7 +233,7 @@ fn cz_split(poly: &F2mPoly, m: u32, irr: &IrreduciblePoly, roots: &mut Vec<F2mEl
         let mut bits = Vec::new();
         let bound = m.min(20);
         for k in 0..bound {
-            if (attempt.wrapping_mul(0x9E3779B1).rotate_left(k as u32) >> 1) & 1 == 1 {
+            if (attempt.wrapping_mul(0x9E3779B1).rotate_left(k) >> 1) & 1 == 1 {
                 bits.push(k);
             }
         }
